backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
distlib==0.3.9
Django==5.1.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
filelock==3.16.1
mysqlclient==2.2.7
packaging==24.2
pipenv==2024.4.0
platformdirs==4.3.6
PyJWT==2.10.1
setuptools==75.6.0
sqlparse==0.5.2
tzdata==2024.2
virtualenv==20.28.0
